requirements.txt
keras==2.10.0
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
scikit_learn==1.4.2
tensorflow_gpu==2.10.0
